pyproject.toml
[tool.poetry]
name = "twitter-dashboard"
version = "0.1.0"
description = ""
authors = ["yokobot <takuou@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.1"
tweepy = "^4.8.0"
DateTime = "^4.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
